requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.3
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
